Bug 912788 - mcMerge: Only use the first line of commit messages; r=graememcc
Only use the first line of the commit message, to avoid false positives when
checking for bug numbers and backouts later. This prevents B2G bumper bot
commits from being detected as a bug in need of marking due to the bug numbers
in the extended commit message. The commit message hook enforces that the bug
number is in the first line, so this won't miss valid cases.