Bug 1072923 - mcMerge: Move handling of B2G repo flags to FlagLoader.js
Prior to this patch, getFlags.php was out of date with respect to the
list of valid B2G release repos. In the future we should fix the
handling of these repos (they use a different versioning scheme and do
not use tracking flags, only status), however to simplify the transition
away from getFlags.php, this patch maintains the unsupported B2G repo
behaviour - that is mcMerge.trackingFlag and mcMerge.statusFlag have
the same value before and after this patch: null.