Bug 1131862 - mcMerge: Display only the active milestones
The bzapi config target_milestone property on each product returns all
milestones for that product, including the inactive ones. A new property has
been added to the API, of form:

     "target_milestone_detail" : [
        {
           "id" : 2336,
           "is_active" : true,
           "name" : "---",
           "sortkey" : 0
        },

We now use target_milestone_detail instead of target_milestone, and filter on
is_active to find only the active milestones. We use map() to flattern the list
of activate milestone names, so existing consumers do not need to be adjusted.