Checkpoint april20
This fixes keyword and comment submissions. I'm seeing issues where the bugherder ui doesn't update correctly (after submitting changes, the submit button is still enabled, and if I go to the summary page, it says nothing was submitted).

Want to get this checked in to my branch so I don't lose the fixes, though.