Bug 1658093 - Add underscore to Thunderbird tracking flag names.
The tracking and status flags for Thunderbird have an underscore in their
names since Thunderbird 30. The resulting flags are cf_tracking_thunderbird_XX
and cf_status_thunderbird_XX. This is different than the assumption that there
is no underscore as is the case for the Firefox flags.
This sticks the additional underscore in the "productName" variable in FlagLoader
as that variable only gets used to generate the flag names.